backend/requirements.txt
asgiref==3.8.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
crispy-bootstrap4==2025.6
cryptography==45.0.6
cycler==0.12.1
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-multiselectfield==1.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.57.0
gunicorn==23.0.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
mysqlclient==2.2.7
numpy==2.2.5
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pygame==2.6.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
